+++ Kexa/services/manageVarEnvironnement.service.ts
@@ -1,6 +1,7 @@
 import axios from "axios";
 
 import {getNewLogger} from "./logger.service";
+import { jsonStringify } from "../helpers/jsonStringify";
 const logger = getNewLogger("KubernetesLogger");
 
 
@@ -146,7 +147,6 @@ async function getEnvVarWithBitwarden(){
 import {listSecrets} from "./addOn/gcpGathering.service";
 import {deleteFile, writeStringToJsonFile} from "../helpers/files";
 import {Storage} from "@google-cloud/storage";
-import { jsonStringify } from "../helpers/jsonStringify";
 async function possibleWithGoogleSecretManager(projectId: any): Promise<boolean> {
     if ((process.env["GOOGLE_APPLICATION_CREDENTIALS"]
         && process.env["GOOGLE_STORAGE_PROJECT_ID"]))