+++ CHANGELOG.md
@@ -1822,3 +1822,34 @@ Kexa/services/addOn.service.ts was changed
 Kexa/services/analyse.service.ts was changed
 
 
+## 1.14.0-devestebancontinuousscan.80.7266ea8
+
+### Files added: 2
+
+Kexa/services/memoisation.service.ts was added
+
+config/demo/exemple6.default.json was added
+
+### Files changed: 10
+
+Kexa/services/memoisation.service.ts was changed
+
+config/demo/exemple6.default.json was changed
+
+Kexa/index.ts was changed
+
+Kexa/main.ts was changed
+
+Kexa/models/settingFile/conditions.models.ts was changed
+
+Kexa/services/addOn.service.ts was changed
+
+Kexa/services/analyse.service.ts was changed
+
+Kexa/services/manageVarEnvironnement.service.ts was changed
+
+Kexa/services/updateCapability.service.ts was changed
+
+documentation/Documentation-Kexa.md was changed
+
+